requirements.txt
fastapi
uvicorn
python-multipart
torch
torchvision
transformers
pillow
accelerate